Use dpms flag with xset
Other seeming improvements given.

xset was not expiring - the screen stayed lit.
Want it to timeout at the seconds given, and
go blank until the (Forth) session types '+++'
(an escape sequence) to wake the screen.

The way this works at present, that means they
may've been typing something 'important' to the
Forth session at the time the screen went blank,
as there's no mechanism (yet?) to keep the session
from blanking due strictly to keyboard activity.

This is because the Xorg system looks for activity
within the Xorg context, and here, that's not the
case (exactly).

While Xorg is displaying the xterm alright, it is
not receiving keystrokes (it can but in practice
it does not receive them .. that honor comes via
the RPi USART, into the Pi, which has converse.py
running to 'hear/listen to' the USART traffic.

This is independent of Xorg (entirely) and can be
done anywhere (say, at the Linux Virtual Console,
entirely outside of Xorg).

So, an artificial means of managing the screen
saver parameters (as xset knows them) while in the
Forth/xterm context, receiving keystrokes via the
Pi's USART.

Best idea so far is this escape business with the
use of '+++' to 'wake up' the session's visual
display on the RPi.

Note that the '+++' sequence requires no video at
the moment it is invoked .. it has no idea that the
current xterm's context has the video blanked.

So it can be typed blind.

-- wa1tnr, 25 November 2019, 16:16 UTC

New branch - xset experiments

	modified:   converse.py

On branch xset_exp-aa-